Restructured code and file tree. Added basic routing
